// src/Entities/Article/Model/consts/articleConsts.ts
// sort only 3 fields
export enum ArticleSortField {
  VIEWS = "views",
  TITLE = "title",
  CREATED = "createdAt",
}
export enum ArticleBlockType {
  CODE = "CODE",
  IMAGE = "IMAGE",
  TEXT = "TEXT",
}
export enum ArticleType {
  ALL = "ALL",
  IT = "IT",
  SCIENCE = "SCIENCE",
  ECONOMICS = "ECONOMICS",
}
export enum ArticleView {
  BIG = "BIG",
  SMALL = "SMALL",
}